+++ .eslintrc.json
@@ -4,11 +4,10 @@
 	"extends": ["prettier", "plugin:react/recommended"],
 	"rules": {
 		"indent": ["error", "tab", { "SwitchCase": 1 }],
-		"react/jsx-no-bind": "error",
+		"react/jsx-no-bind": "warn",
 		"react/require-default-props": "error",
 		"react/forbid-component-props": ["error", { "forbid": ["style"] }],
 		"react/forbid-dom-props": ["error", { "forbid": ["style"] }],
-		"react/no-danger": "error",
 		"react/no-this-in-sfc": "error",
 		"react/no-unused-prop-types": "warn",
 		"react/no-unused-state": "warn",
