--- src/components/Selector/index.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-class Selector extends React.Component {
-	render() {
-		let componentClass = 'selector';
-		if (this.props.activeID === this.props.id) {
-			componentClass = 'selector active';
-		}
-		return <div className={componentClass} onClick={() => this.props.onHandleClick()} />;
-	}
-}
-
-export default Selector;
