--- src/components/dashboard/dashboard.css
@@ -1,3 +1,0 @@
-.dashboard{
-	margin-top: 7%;
-}
