--- src/components/footer/styles.css
@@ -1,134 +1,0 @@
-.links-section {
-  display: flex;
-  flex-direction: column;
-}
-
-.links-section {
-  width: fit-content;
-}
-
-.fest-logo {
-  width: 25vw;
-}
-
-
-.festnew-logo {
-  width: 100%;
-  padding-left: 4%;
-  padding-top: 8%;
-  display: none;
-}
-
-.foot-int {
-  width: 50%;
-  margin-left: 13%;
-  position: relative;
-  left: -6%;
-  top: 10px;
-  height: 100%;
-}
-
-.foot-fest-logo{
-  margin-left: 30%;
-  width:80%;
-}
-
-.foot-int-logo{
-  margin-left: 30%;
-  max-width: 130%;
-}
-
-.festpic {
-  width: 35%;
-  height: 100%;
-}
-
-.white-link {
-  color: antiquewhite;
-}
-
-.white-link:hover {
-  border-bottom: 1px dashed #FF6E2C;
-  text-decoration: none;
-  color: white;
-}
-
-.footer {
-  display: flex;
-  justify-content: space-between;
-}
-
-.contact-section {
-  width: 40vw;
-}
-
-.footer-links {
-  border-bottom: 1px dashed #FF6E2C;
-  margin-bottom: 4px;
-  width: fit-content;
-  color: white;
-}
-
-.section-title {
-  color: white;
-  padding-left: 8px;
-  border-left: 4px solid #FF6E2C;
-}
-
-.icons {
-  display: flex;
-  margin-bottom: 20px;
-}
-
-.single-icon {
-  size: 0.5rem;
-  padding-right: 1rem;
-  font-size: x-large;
-  color: #FF6E2C;
-  padding-top:0px;
-}
-
-.single-icon:hover {
-  size: 1rem;
-  font-size: x-large;
-  color: whitesmoke;
-  padding-top:0px;
-}
-
-.links-section a {
-  text-decoration: none;
-}
-
-@media screen and (max-width: 560px) and (min-width: 320px) {
-  .uk-card.uk-card-body.uk-margin-left {
-    left:-1%;
-    width: 100%;
-    position: relative;
-    top: 0%;
-    font-size: 10px;
-  }
-
-  .fest-logo {
-    width: 40vw;
-    position: relative;
-    display: none;
-  }
-
-  .icons {
-    display: -ms-flexbox;
-    display: flex;
-    position: relative;
-    left: -5%;
-    top: -5%;
-  }
-
-  .footer {
-    display: flex;
-    justify-content: center;
-  }
-
-  .festnew-logo {
-    display: flex !important;
-  }
-
-}
