--- src/components/header/styles.css
@@ -1,58 +1,0 @@
-#header {
-  position: relative;
-  top: 0px;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: 500;
-  background-color: black;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.heading-text {
-  color: white !important;
-}
-
-.heading-text span {
-  padding-bottom: 6px;
-}
-
-.uk-active .heading-text span {
-  border-bottom: 2px solid #FF6E2C;
-  padding-bottom: 4px;
-}
-
-
-.institute-logo {
-  width: 30%;
-  margin-top: 0.5em;
-  margin-left: 0.5%;
-  visibility: hidden;
-}
-@media screen and (max-width: 560px) and (min-width: 320px) {
-  .uk-navbar {
-    display: -ms-flexbox;
-    display: flex;
-    position: relative;
-    left: 13%;
-}
-  .uk-navbar-center:not(:only-child) {
-    position: absolute;
-    top: 120%;
-    left: 44%;
-  }
-  .firebaseui-idp-button.mdl-button.mdl-js-button.mdl-button--raised.firebaseui-idp-google.firebaseui-id-idp-button {
-
-    position: relative !important;
-    left: 0%;
-  }
-  .countdown{
-    height: calc(100vh - 175px);
-  }
-  .uk-navbar-item, .uk-navbar-nav > li > a, .uk-navbar-toggle {
-
-    padding: 0px 9px;
-
-  }
-}
-
