--- src/components/team/team.css
@@ -1,11 +1,0 @@
-.separator{
-	width: 60%;
-	margin: 5% auto;
-	color: #ff6e2c;
-	opacity: 0.3;
-	padding-top: 2%;
-}
-
-.team-section{
-	padding: 5px;
-}
