--- src/components/timeline/timeline.css
@@ -1,66 +1,0 @@
-
-
-.vertical-timeline-element-content .vertical-timeline-element-date, .vertical-timeline-element-content {
-  color: #FF6E2C;
-  font-family: Vidaloka;
-}
-
-.vertical-timeline-element-content p {
-  color: #696969;
-}
-
-.vertical-timeline--animate .vertical-timeline-element-content.bounce-in {
-  color: white;
-}
-
-.vertical-timeline-element-content::before {
-  border-left-color: #FF6E2C;
-}
-
-.vertical-timeline-element.vertical-timeline-element--right .vertical-timeline-element-content::before, .vertical-timeline-element:nth-child(2n):not(.vertical-timeline-element--left) .vertical-timeline-element-content::before {
-  border-right-color: #FF6E2C;
-}
-
-.vertical-timeline-element-date {
-  color: #FF6E2C;
-}
-
-.tl-title{
-  color: white;
-  text-align: center;
-  margin-top: 40px;
-  margin-bottom: 5px;
-  text-decoration: underline;
-  text-underline-position: under;
-  text-decoration-color: #FF6E2C;
-  font-family: Vidaloka;
-}
-
-/*.vertical-timeline-element-icon{
-	width: 20px;
-	height: 20px;
-	top: 10px;
-    position: absolute;
-    left: 2
-	border-radius: 50%;
-	color: #ff6e2c;
-}*/
-@media only screen and (min-width: 1170px)
-
-{
-.vertical-timeline-element-icon {
-    width: 29px !important;
-    height: 29px !important;
-    left: 51.4% !important;
-    margin-left: -30px;
-    -webkit-transform: translateZ(0);
-    -webkit-backface-visibility: hidden;
-}
-}
-
-@media screen and (max-width: 560px) and (min-width: 320px) {
-
-.vertical-timeline-element-content::before {
-    border-left-color: transparent;
-}
-}
