--- src/css/index.css
@@ -1,155 +1,0 @@
-
-@font-face{
-  font-family: vidaloka;
-  src:url(https://fonts.googleapis.com/css?family=Vidaloka);
-}
-
-html, body {
-  background: black;
-  height: 100%;
-  overflow-x: hidden!important;
-  color: white;
-  font-family: 'Roboto Mono', monospace;
-}
-@media screen and (max-width: 560px) and (min-width: 320px) {
- html{
-   font-size: 15px !important;
- }
-}
-.team-ext, .team-ext:hover, .team-ext:focus {
-  color: #FF6E2C;
-}
-
-.timer {
-  display: flex;
-  color: #FF6E2C;
-}
-
-.countdown-timer {
-  align-items: flex-start;
-}
-
-.team {
-  display: flex;
-  justify-content: space-around;
-}
-
-.team-card {
-  width: 300px;
-}
-
-.user-img {
-  height: 100px;
-  width: 100px;
-}
-
-.user-name {
-  margin-left: 10px;
-}
-
-.update-mobile {
-  display: flex;
-}
-
-.dashboard-user-content {
-  display: flex;
-  width: 100vw;
-  justify-content: space-between;
-  flex-wrap: wrap;
-}
-
-.primary-color {
-  color: #FF6E2C;
-  padding-bottom: 5px;
-}
-
-.primary-color-border {
-  border-color: #FF6E2C;
-  background-color: #fff0;
-}
-
-.white {
-  color: white;
-}
-
-.event-description {
-  width: 50%;
-}
-
-.vertical-timeline-element-content .vertical-timeline-element-date {
-  color: #FF6E2C;
-}
-
-.vertical-timeline--animate .vertical-timeline-element-content.bounce-in {
-  background-color: transparent;
-  -webkit-box-shadow: 0 0 0 0 transparent;
-  -moz-box-shadow: 0 0 0 0 transparent;
-  box-shadow: 0 0 0 0 transparent;
-  border: 1px dashed #FF6E2C;
-  color: white;
-  opacity: 1;
-  cursor: pointer;
-}
-
-.vertical-timeline-element-title {
-  font-size: 25px;
-}
-
-.absolute {
-  position: absolute;
-}
-
-.square {
-  top: 0;
-  left: 0;
-  height: 200px;
-  width: 200px;
-  background: url('https://firebasestorage.googleapis.com/v0/b/cerebro-2018-f1052.appspot.com/o/square.png?alt=media&token=8979df90-1eaf-466e-9ef9-94546eec3e90') no-repeat;
-}
-
-.triangle {
-  top: 0;
-  right: 0;
-  height: 200px;
-  width: 200px;
-  background: url('https://firebasestorage.googleapis.com/v0/b/cerebro-2018-f1052.appspot.com/o/triangle.png?alt=media&token=f3cbbbed-2d31-4201-a051-8d123d594156') no-repeat;
-}
-
-canvas {
-  margin-top: -100vh;
-}
-
-.firebaseui-card-content {
-  padding: 0;
-}
-
-.firebaseui-idp-google, .firebaseui-idp-google:hover, .mdl-button.firebaseui-idp-google:active, .mdl-button.firebaseui-idp-google:focus {
-  background-color: transparent;
-}
-
-.firebaseui-idp-google>.firebaseui-idp-text {
-  color: white;
-  visibility: hidden;
-}
-
-.firebaseui-idp-google>.firebaseui-idp-text:after {
-  color: white;
-  visibility: visible;
-  content: 'SIGN IN';
-  display: block;
-  margin-top: -15px;
-  margin-left: -7px;
-}
-@media screen and (max-width: 560px) and (min-width: 320px) {
-  .firebaseui-idp-google>.firebaseui-idp-text:after {
-  margin-top: -30px;
-  }
-}
-
-div#root {
-    border: 3px solid #ff6e2c;
-}
-
-.fp-tableCell {
-  vertical-align: top!important;
-}
