+++ src/reducers/userPhoneReducer.js
@@ -1,11 +1,16 @@
-import { actionTypes } from '../actions/userPhoneReducer';
+import { actionTypes } from '../actions/userPhoneActions';
 
-const initialState = { phone: null };
+const initialState = {
+	phone: null
+};
 
 export const userPhoneReducer = (state = initialState, action) => {
 	switch (action.type) {
 		case actionTypes.LOAD_PHONE:
-			return { ...state, phone: action.phone };
+			return {
+				...state,
+				phone: action.phone
+			};
 		default:
 			return state;
 	}
