+++ src/styles/About.css
@@ -1,0 +1,15 @@
+.about-title {
+	padding-left: 8px;
+	border-left: 4px solid #ff6e2c;
+	font-size: xx-large;
+}
+
+@media screen and (min-width: 1120px) {
+	.uk-width-1-5\@l {
+		width: 25% !important;
+	}
+	.uk-container {
+		padding-left: 15% !important;
+		padding-right: 15% !important;
+	}
+}
