+++ src/styles/Countdown.css
@@ -1,0 +1,46 @@
+.countdown {
+	display: flex;
+	height: calc(100vh - 72px);
+	align-items: center;
+	justify-content: center;
+	flex-direction: column;
+	font-family: Vidaloka;
+}
+
+.cerebro-hero {
+	height: 200px;
+}
+
+.uk-countdown-number {
+	font-size: 35px;
+	color: #ff6e2c;
+}
+
+.uk-countdown-number {
+	font-size: 35px;
+}
+
+.countdown-header {
+	margin-top: -35px;
+	padding: 4px 16px;
+	background-color: black;
+}
+
+@media screen and (max-width: 560px) and (min-width: 320px) {
+	.firebaseui-idp-button.mdl-button.mdl-js-button.mdl-button--raised.firebaseui-idp-google.firebaseui-id-idp-button {
+		position: fixed;
+		left: 95%;
+	}
+	.countdown {
+		height: calc(100vh - 175px);
+	}
+	.cerebro-hero {
+		height: auto;
+		width: 275px;
+	}
+}
+
+.next-year {
+	font-family: 'Roboto Mono', monospace;
+	font-size: 20px;
+}
