+++ src/styles/Dashboard.css
@@ -1,0 +1,3 @@
+.dashboard {
+	margin-top: 7%;
+}
