+++ src/styles/EventList.css
@@ -1,0 +1,100 @@
+@import url('https://fonts.googleapis.com/css?family=Slabo+27px');
+
+.event-list-heading {
+	font-size: 25px;
+	margin-bottom: 5px;
+	border-bottom: 2px solid #ff6e2c;
+}
+
+.start-button {
+	margin-top: 30px;
+	font-size: 12px;
+	font-family: 'Roboto Mono', monospace;
+	z-index: 1;
+}
+
+.start-button:hover {
+	cursor: pointer;
+}
+
+.timer {
+	font-size: 25px;
+}
+
+.uk-grid-small > * {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+}
+
+.fp-section {
+	padding-right: 5% !important;
+	padding-left: 5% !important;
+}
+
+.fp-tableCell {
+	vertical-align: top;
+	padding: 30px 5% 0;
+}
+
+.uk-slider-items li {
+	width: 33.33%;
+}
+
+.full-width {
+	width: 100%;
+}
+
+#slider-arrow-left,
+#slider-arrow-right {
+	height: 35px;
+	width: 35px;
+	top: 37%;
+	font-size: 100px;
+	color: grey;
+	transition: 0.1s ease-in-out;
+}
+
+#slider-arrow-left:hover,
+#slider-arrow-right:hover {
+	color: #ff6e2c;
+}
+
+#slider-arrow-left {
+	left: 50px;
+}
+
+#slider-arrow-right {
+	right: 50px;
+}
+
+/* Medium devices */
+@media screen and (max-width: 959px) {
+	.uk-slider-items li {
+		width: 50%;
+	}
+
+	#slider-arrow-left {
+		left: 1%;
+	}
+
+	#slider-arrow-right {
+		right: 1%;
+	}
+}
+
+/* Small devices */
+@media screen and (max-width: 640px) {
+	.uk-slider-items li {
+		width: 99%;
+	}
+
+	.uk-dotnav > * {
+		padding-left: 7px;
+	}
+
+	#slider-arrow-left,
+	#slider-arrow-right {
+		display: none;
+	}
+}
