+++ src/styles/Panel.css
@@ -1,0 +1,90 @@
+.event-panel {
+	border: 0.1rem solid rgba(255, 255, 255, 0.22);
+	-webkit-transition: all 0.3s;
+	-moz-transition: all 0.3s;
+	-ms-transition: all 0.3s;
+	-o-transition: all 0.3s;
+	transition: all 0.3s;
+	user-select: none;
+	-moz-user-select: none;
+	-ms-user-select: none;
+	-webkit-user-select: none;
+}
+
+.event-panel:hover {
+	border: 0.1rem solid #ff6e2c;
+	-webkit-transition: all 0.3s;
+	-moz-transition: all 0.3s;
+	-ms-transition: all 0.3s;
+	-o-transition: all 0.3s;
+	transition: all 0.3s;
+}
+
+.panel-content {
+	height: 6em;
+	overflow: hidden;
+}
+
+.panel-team-size {
+	font-family: monospace;
+	font-size: 12px;
+}
+
+.panel-register {
+	background-color: #ff6e2c;
+	padding: 9px 17px;
+	cursor: pointer;
+	white-space: nowrap;
+	margin-left: -19px;
+}
+
+.uk-card-title {
+	font-size: 1.4rem;
+	line-height: 1.2;
+	height: 2.4em;
+	overflow: hidden;
+}
+
+.panel-unregister {
+	background-color: white;
+	padding: 8px 8px;
+	cursor: pointer;
+	white-space: nowrap;
+	color: #ff6e2c;
+	margin-left: -20px;
+}
+
+.panel-more {
+	border: 0.1rem solid #ff6e2c;
+	padding: 0.5rem 0.5rem;
+	cursor: pointer;
+	white-space: nowrap;
+	margin-left: 8px;
+}
+
+.panel-more:hover {
+	border: 0.1rem solid white;
+	background-color: rgb(246, 104, 38);
+	cursor: pointer;
+	white-space: nowrap;
+	margin-left: 8px;
+}
+
+.panel-buttons {
+	display: flex;
+	justify-content: space-between;
+}
+
+.uk-card.uk-card-body.event-panel {
+	padding: 30px 30px !important;
+}
+
+.panel-img-container {
+	display: flex;
+	justify-content: center;
+	height: 91px;
+}
+
+.panel-img {
+	object-fit: contain;
+}
