+++ src/styles/Team.css
@@ -1,0 +1,18 @@
+.separator {
+	width: 60%;
+	margin: 5% auto;
+	color: #ff6e2c;
+	opacity: 0.3;
+	padding-top: 2%;
+}
+
+.team-section {
+	padding: 5px;
+}
+
+.fest-logo {
+	width: '100%';
+	overflow: 'hidden';
+	background-size: 'cover';
+	background-position: 'center';
+}
